Moved triggers execution from case template creation to case creation
